requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2
pillow==11.2.1
psycopg2-binary==2.9.10
pytz==2025.2
reportlab==4.4.0
sqlparse==0.5.3
tzdata==2025.2
